requirements_small.txt
fastapi[all]
faster-whisper
openai
pandas
python-dotenv
redis